refactor(test): rename setup repo module to test repo
